requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Django==4.2.11
django-bootstrap-v5==1.0.11
django-cors-headers==4.3.1
django-phonenumber-field==7.3.0
django-seed==0.3.1
django-tailwind==3.8.0
djangorestframework==3.15.1
Faker==24.11.0
mysql==0.0.3
mysqlclient==2.2.4
pillow==10.3.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
toposort==1.10
tzdata==2024.1
